[Brave News] Move configure dialog into brave_news resources
