+++ components/brave_news/browser/resources/customize/ChannelCard.tsx
@@ -6,10 +6,10 @@
 import Flex from '$web-common/Flex'
 import * as React from 'react'
 import styled from 'styled-components'
-import { useChannelSubscribed } from '../../../../../brave_news/browser/resources/shared/Context'
-import { channelIcons } from '../../../../../brave_news/browser/resources/shared/Icons'
-import FollowButton from '../../../../../brave_news/browser/resources/shared/FollowButton'
-import { getTranslatedChannelName } from '../../../../../brave_news/browser/resources/shared/channel'
+import { useChannelSubscribed } from '../shared/Context'
+import { channelIcons } from '../shared/Icons'
+import FollowButton from '../shared/FollowButton'
+import { getTranslatedChannelName } from '../shared/channel'
 
 const SubscribeButton = styled(FollowButton)`
     position: absolute;
