+++ components/brave_news/browser/resources/customize/Discover.tsx
@@ -10,10 +10,10 @@ import Button from '@brave/leo/react/button'
 import * as React from 'react'
 import { useState } from 'react'
 import styled from 'styled-components'
-import { useBraveNews, useChannels } from '../../../../../brave_news/browser/resources/shared/Context'
+import { useBraveNews, useChannels } from '../shared/Context'
 import ChannelCard from './ChannelCard'
 import DiscoverSection from './DiscoverSection'
-import PublisherCard, { DirectPublisherCard } from '../../../../../brave_news/browser/resources/shared/PublisherCard'
+import PublisherCard, { DirectPublisherCard } from '../shared/PublisherCard'
 import { PopularCarousel } from './Popular'
 import { SuggestionsCarousel } from './Suggestions'
 import useSearch from './useSearch'
