+++ components/brave_news/browser/resources/customize/DiscoverSection.tsx
