+++ components/brave_news/browser/resources/customize/colors.ts
