+++ components/brave_news/browser/resources/customize/useSearch.ts
@@ -4,8 +4,8 @@
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
 import * as React from 'react'
-import { useBraveNews, useChannels } from '../../../../../brave_news/browser/resources/shared/Context'
-import getBraveNewsController, { FeedSearchResultItem } from '../../../../../brave_news/browser/resources/shared/api'
+import { useBraveNews, useChannels } from '../shared/Context'
+import getBraveNewsController, { FeedSearchResultItem } from '../shared/api'
 
 export const QUERY_MIN_LENGTH_FILTER_SOURCES = 2
