+++ components/brave_news/browser/resources/shared/PublisherCard.tsx
@@ -7,7 +7,7 @@ import Flex from '$web-common/Flex'
 import * as React from 'react'
 import { useState } from 'react'
 import styled from 'styled-components'
-import { getCardColor } from '../../../../brave_new_tab_ui/components/default/braveNews/customize/colors'
+import { getCardColor } from '../customize/colors'
 import { usePublisher, usePublisherFollowed } from './Context'
 import FollowButton from './FollowButton'
 import getBraveNewsController from './api'
